+++ MWSE DLL/source/TES3MACHINE.h
@@ -64,6 +64,7 @@ struct TES3MACHINE : public VIRTUALMACHINE
 		long const index);
 	long SetArrayValue(std::string const& caller, long const id,
 		long const index, long const value);
+	long GetArraySize(std::string const& caller, long const id);
 	std::vector<std::vector<long> >& arrays();
 
 private: